Raspberry/RFID/requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
lgpio==0.2.2.0
mfrc522==0.0.7
requests==2.32.5
rpi-gpio-emu==0.0.1
rpi-lgpio==0.6
spidev==3.8
urllib3==2.6.3
